requirements.txt
pdfkit
PyJWT
django-rest-auth
django_allauth==0.45.0
django_filter==21.1
six==1.16.0
Django==3.2
django_resized==0.3.11
dj_rest_auth==2.1.11
django_environ==0.4.5
requests==2.26.0
djangorestframework==3.12.4
python-decouple==3.5
python_dateutil==2.8.2
djangorestframework_simplejwt
django-crispy-forms
django-ckeditor
django-cors-headers
pillow
django-seed
